+++ code/modules/mob/living/living.dm
@@ -3,6 +3,7 @@
 
 	//Prime this list if we need it.
 	if(has_huds)
+		// Note, this should be refactored to drop priority overlays
 		add_overlay(backplane,TRUE) //Strap this on here, to block HUDs from appearing in rightclick menus: http://www.byond.com/forum/?post=2336679
 		hud_list = list()
 		hud_list.len = TOTAL_HUDS
@@ -18,10 +19,13 @@
 
 	selected_image = image(icon = GLOB.buildmode_hud, loc = src, icon_state = "ai_sel")
 
+	AddElement(/datum/element/spontaneous_vore)
+
 /mob/living/proc/get_visible_name()
-	var/datum/component/shadekin/SK = get_shadekin_component()
-	if(SK && SK.in_phase)
-		return "Something"
+	var/list/name_data = list(null)
+	if(SEND_SIGNAL(src, COMSIG_HUMAN_GET_VISIBLE_NAME, name_data) & COMPONENT_VISIBLE_NAME_CHANGED)
+		return name_data[1]
+
 	if(real_name)
 		return real_name
 	else
@@ -75,12 +79,7 @@
 			tf_mob_holder.forceMove(get_dat_turf)
 		QDEL_LIST_NULL(tf_mob_holder.vore_organs)
 		tf_mob_holder.vore_organs = list()
-		for(var/obj/belly/B as anything in vore_organs)
-			B.loc = tf_mob_holder
-			B.forceMove(tf_mob_holder)
-			B.owner = tf_mob_holder
-			tf_mob_holder.vore_organs |= B
-			vore_organs -= B
+		tf_mob_holder.mob_belly_transfer(src)
 	if(tf_mob_holder)
 		tf_mob_holder = null
 	QDEL_NULL_LIST(hud_list)
@@ -95,14 +94,15 @@
 			organs -= OR
 			qdel(OR)
 
-	if(LAZYLEN(internal_organs) && !istype(src, /mob/living/simple_mob/animal))
+	if(LAZYLEN(internal_organs))
 		internal_organs_by_name.Cut()
 		while(internal_organs.len)
 			var/obj/item/OR = internal_organs[1]
 			internal_organs -= OR
-			qdel(OR)
+			if(isobj(OR))
+				qdel(OR)
 
-	cultnet.updateVisibility(src, 0)
+	GLOB.cultnet.updateVisibility(src, 0)
 
 	if(aiming)
 		qdel(aiming)
@@ -116,7 +116,7 @@
 	set name = "Pull"
 	set category = "Object"
 
-	if(AM.Adjacent(src))
+	if(istype(AM) && AM.Adjacent(src))
 		src.start_pulling(AM)
 
 	return
@@ -170,7 +170,7 @@
 		manual_afk = TRUE
 
 /mob/living/proc/updatehealth()
-	if(SEND_SIGNAL(src, COMSIG_UPDATE_HEALTH) & COMSIG_UPDATE_HEALTH_GOD_MODE)
+	if(SEND_SIGNAL(src, COMSIG_LIVING_HEALTH_UPDATE) & COMSIG_LIVING_HEALTH_UPDATE_GOD_MODE)
 		health = getMaxHealth()
 		set_stat(CONSCIOUS)
 	else
@@ -467,6 +467,7 @@
 	return result
 
 ///Use this proc to get the damage in which the mob will be put into critical condition (hardcrit)
+///Will return a NEGATIVE value. Ex: MaxHealth of 100 returns -50
 /mob/living/proc/get_crit_point()
 	return -(getMaxHealth()*0.5)
 
@@ -490,7 +491,7 @@
 		cloneloss = round(cloneloss / h_mult)
 	maxHealth = newMaxHealth
 
-/mob/living/Stun(amount)
+/mob/living/Stun(amount, ignore_canstun = FALSE)
 	for(var/datum/modifier/M in modifiers)
 		if(!isnull(M.disable_duration_percent))
 			amount = round(amount * M.disable_duration_percent)
@@ -498,14 +499,14 @@
 	if(stunned > 0)
 		add_status_indicator("stunned")
 
-/mob/living/SetStunned(amount)
+/mob/living/SetStunned(amount, ignore_canstun = FALSE)
 	..()
 	if(stunned <= 0)
 		remove_status_indicator("stunned")
 	else
 		add_status_indicator("stunned")
 
-/mob/living/AdjustStunned(amount)
+/mob/living/AdjustStunned(amount, ignore_canstun = FALSE)
 	if(amount > 0)
 		for(var/datum/modifier/M in modifiers)
 			if(!isnull(M.disable_duration_percent))
@@ -516,7 +517,7 @@
 	else
 		add_status_indicator("stunned")
 
-/mob/living/Weaken(amount)
+/mob/living/Weaken(amount, ignore_canstun = FALSE)
 	for(var/datum/modifier/M in modifiers)
 		if(!isnull(M.disable_duration_percent))
 			amount = round(amount * M.disable_duration_percent)
@@ -524,14 +525,14 @@
 	if(weakened > 0)
 		add_status_indicator("weakened")
 
-/mob/living/SetWeakened(amount)
+/mob/living/SetWeakened(amount, ignore_canstun = FALSE)
 	..()
 	if(weakened <= 0)
 		remove_status_indicator("weakened")
 	else
 		add_status_indicator("weakened")
 
-/mob/living/AdjustWeakened(amount)
+/mob/living/AdjustWeakened(amount, ignore_canstun = FALSE)
 	if(amount > 0)
 		for(var/datum/modifier/M in modifiers)
 			if(!isnull(M.disable_duration_percent))
@@ -542,7 +543,7 @@
 	else
 		add_status_indicator("weakened")
 
-/mob/living/Paralyse(amount)
+/mob/living/Paralyse(amount, ignore_canstun = FALSE)
 	for(var/datum/modifier/M in modifiers)
 		if(!isnull(M.disable_duration_percent))
 			amount = round(amount * M.disable_duration_percent)
@@ -550,14 +551,14 @@
 	if(paralysis > 0)
 		add_status_indicator("paralysis")
 
-/mob/living/SetParalysis(amount)
+/mob/living/SetParalysis(amount, ignore_canstun = FALSE)
 	..()
 	if(paralysis <= 0)
 		remove_status_indicator("paralysis")
 	else
 		add_status_indicator("paralysis")
 
-/mob/living/AdjustParalysis(amount)
+/mob/living/AdjustParalysis(amount, ignore_canstun = FALSE)
 	if(amount > 0)
 		for(var/datum/modifier/M in modifiers)
 			if(!isnull(M.disable_duration_percent))
@@ -568,7 +569,7 @@
 	else
 		add_status_indicator("paralysis")
 
-/mob/living/Sleeping(amount)
+/mob/living/Sleeping(amount, ignore_canstun = FALSE)
 	for(var/datum/modifier/M in modifiers)
 		if(!isnull(M.disable_duration_percent))
 			amount = round(amount * M.disable_duration_percent)
@@ -576,14 +577,14 @@
 	if(sleeping > 0)
 		add_status_indicator("sleeping")
 
-/mob/living/SetSleeping(amount)
+/mob/living/SetSleeping(amount, ignore_canstun = FALSE)
 	..()
 	if(sleeping <= 0)
 		remove_status_indicator("sleeping")
 	else
 		add_status_indicator("sleeping")
 
-/mob/living/AdjustSleeping(amount)
+/mob/living/AdjustSleeping(amount, ignore_canstun = FALSE)
 	if(amount > 0)
 		for(var/datum/modifier/M in modifiers)
 			if(!isnull(M.disable_duration_percent))
@@ -594,7 +595,7 @@
 	else
 		add_status_indicator("sleeping")
 
-/mob/living/Confuse(amount)
+/mob/living/Confuse(amount, ignore_canstun = FALSE)
 	for(var/datum/modifier/M in modifiers)
 		if(!isnull(M.disable_duration_percent))
 			amount = round(amount * M.disable_duration_percent)
@@ -602,14 +603,14 @@
 	if(confused > 0)
 		add_status_indicator("confused")
 
-/mob/living/SetConfused(amount)
+/mob/living/SetConfused(amount, ignore_canstun = FALSE)
 	..()
 	if(confused <= 0)
 		remove_status_indicator("confused")
 	else
 		add_status_indicator("confused")
 
-/mob/living/AdjustConfused(amount)
+/mob/living/AdjustConfused(amount, ignore_canstun = FALSE)
 	if(amount > 0)
 		for(var/datum/modifier/M in modifiers)
 			if(!isnull(M.disable_duration_percent))
@@ -620,7 +621,7 @@
 	else
 		add_status_indicator("confused")
 
-/mob/living/Blind(amount)
+/mob/living/Blind(amount, ignore_canstun = FALSE)
 	for(var/datum/modifier/M in modifiers)
 		if(!isnull(M.disable_duration_percent))
 			amount = round(amount * M.disable_duration_percent)
@@ -628,14 +629,14 @@
 	if(eye_blind > 0)
 		add_status_indicator("blinded")
 
-/mob/living/SetBlinded(amount)
+/mob/living/SetBlinded(amount, ignore_canstun = FALSE)
 	..()
 	if(eye_blind <= 0)
 		remove_status_indicator("blinded")
 	else
 		add_status_indicator("blinded")
 
-/mob/living/AdjustBlinded(amount)
+/mob/living/AdjustBlinded(amount, ignore_canstun = FALSE)
 	if(amount > 0)
 		for(var/datum/modifier/M in modifiers)
 			if(!isnull(M.disable_duration_percent))
@@ -725,12 +726,11 @@
 	BITSET(hud_updateflag, HEALTH_HUD)
 	BITSET(hud_updateflag, STATUS_HUD)
 	BITSET(hud_updateflag, LIFE_HUD)
-	ExtinguishMob()
-	fire_stacks = 0
 	if(ai_holder) // AI gets told to sleep when killed. Since they're not dead anymore, wake it up.
 		ai_holder.go_wake()
 
 	SEND_SIGNAL(src, COMSIG_HUMAN_DNA_FINALIZED)
+	SEND_SIGNAL(src, COMSIG_LIVING_AHEAL)
 
 /mob/living/proc/rejuvenate()
 	if(reagents)
@@ -823,7 +823,7 @@
 	set name = "Resist"
 	set category = "IC.Game"
 
-	if(!incapacitated(INCAPACITATION_KNOCKOUT) && (last_resist_time + RESIST_COOLDOWN < world.time))
+	if(!incapacitated(INCAPACITATION_KNOCKOUT) && !is_paralyzed() && (last_resist_time + RESIST_COOLDOWN < world.time))
 		last_resist_time = world.time
 		resist_grab()
 		if(!weakened)
@@ -854,8 +854,6 @@
 			resist_fire() //stop, drop, and roll
 		else
 			resist_restraints()
-
-	if(attempt_vr(src,"vore_process_resist",args)) return TRUE
 
 /mob/living/proc/resist_buckle()
 	if(buckled)
@@ -888,7 +886,7 @@
 	update_canmove()
 
 //called when the mob receives a bright flash
-/mob/living/flash_eyes(intensity = FLASH_PROTECTION_MODERATE, override_blindness_check = FALSE, affect_silicon = FALSE, visual = FALSE, type = /obj/screen/fullscreen/flash)
+/mob/living/flash_eyes(intensity = FLASH_PROTECTION_MODERATE, override_blindness_check = FALSE, affect_silicon = FALSE, visual = FALSE, type = /atom/movable/screen/fullscreen/flash)
 	if(override_blindness_check || !(disabilities & BLIND))
 		overlay_fullscreen("flash", type)
 		spawn(25)
@@ -906,6 +904,9 @@
 
 /mob/living/proc/has_eyes()
 	return 1
+
+/mob/living/proc/has_lungs()
+	return TRUE
 
 /mob/living/proc/get_restraining_bolt()
 	var/obj/item/implant/restrainingbolt/RB = locate() in src
@@ -1016,6 +1017,11 @@
 			else
 				puke_bucket.reagents.add_reagent(REAGENT_ID_TOXIN, rand(3, 6))
 		distance = 0
+	else if(isbelly(loc))
+		var/obj/belly/belly = loc
+		if(message)
+			to_chat(src, span_bolddanger("You throw up inside [belly.owner]'s [belly]!"))
+		distance = 0
 	else
 		if(message)
 			visible_message(span_danger("[src] throws up!"), span_userdanger("You throw up!"))
@@ -1080,40 +1086,44 @@
 
 /mob/living/update_canmove()
 	if(!resting && cannot_stand() && can_stand_overridden())
-		lying = 0
-		canmove = 1
+		lying = FALSE
+		canmove = TRUE
 	else
 		if(istype(buckled, /obj/vehicle))
 			var/obj/vehicle/V = buckled
 			if(is_physically_disabled())
-				lying = 0
-				canmove = 1
+				lying = FALSE
+				canmove = TRUE
 				if(!V.riding_datum) // If it has a riding datum, the datum handles moving the pixel_ vars.
 					pixel_y = V.mob_offset_y - 5
 			else
 				if(buckled.buckle_lying != -1)
 					lying = buckled.buckle_lying
-				canmove = 1
+				canmove = TRUE
 				if(!V.riding_datum) // If it has a riding datum, the datum handles moving the pixel_ vars.
 					pixel_y = V.mob_offset_y
 		else if(buckled)
 			anchored = TRUE
-			canmove = 1 //The line above already makes the chair not swooce away if the sitter presses a button. No need to incapacitate them as a criminally large amount of mechanics read this var as a type of stun.
+			canmove = TRUE //The line above already makes the chair not swooce away if the sitter presses a button. No need to incapacitate them as a criminally large amount of mechanics read this var as a type of stun.
 			if(istype(buckled))
 				if(buckled.buckle_lying != -1)
 					lying = buckled.buckle_lying
 					canmove = buckled.buckle_movable
 				if(buckled.buckle_movable)
 					anchored = FALSE
-					canmove = 1
+					canmove = TRUE
 		else
 			lying = incapacitated(INCAPACITATION_KNOCKDOWN)
 			canmove = !incapacitated(INCAPACITATION_DISABLED)
 
 	if(incapacitated(INCAPACITATION_KNOCKOUT) || incapacitated(INCAPACITATION_STUNNED)) // Making sure we're in good condition to crawl
-		canmove = 0
+		canmove = FALSE
 	else
-		canmove = 1
+		canmove = TRUE
+
+	if(is_paralyzed())
+		lying = TRUE
+		canmove = FALSE
 
 	if(lying)
 		density = FALSE
@@ -1172,10 +1182,6 @@
 	if(isSynthetic())
 		return FALSE
 	return TRUE
-
-// Gets the correct icon_state for being on fire. See OnFire.dmi for the icons.
-/mob/living/proc/get_fire_icon_state()
-	return "generic"
 
 // Called by job_controller.
 /mob/living/proc/equip_post_job()
@@ -1269,8 +1275,6 @@
 	// We just swapped hands, so the thing in our inactive hand will notice it's not the focus
 	var/obj/item/I = get_inactive_hand()
 	if(I)
-		if(I.zoom)
-			I.zoom()
 		I.in_inactive_hand(src)	//This'll do specific things, determined by the item
 	return
 
@@ -1288,7 +1292,7 @@
 		swap_hand()
 
 /mob/living/throw_item(atom/target)
-	if(incapacitated() || !target || istype(target, /obj/screen) || is_incorporeal())
+	if(incapacitated() || !target || istype(target, /atom/movable/screen) || is_incorporeal())
 		return FALSE
 
 	var/atom/movable/item = src.get_active_hand()
@@ -1321,7 +1325,6 @@
 				src.inertia_dir = get_dir(target, src)
 				step(src, inertia_dir)
 			item.throw_at(target, throw_range, item.throw_speed, src)
-			item.throwing = 1 //Small edit so thrown interactions actually work!
 			return TRUE
 		else
 			return FALSE
@@ -1414,6 +1417,9 @@
 /mob/living/proc/adjust_nutrition(amount)
 	nutrition = between(0, nutrition + amount, max_nutrition)
 
+/mob/living/proc/nutrition_percent()
+	return 100 * nutrition / max_nutrition
+
 /mob/living/vv_get_header()
 	. = ..()
 	var/refid = REF(src)
@@ -1434,7 +1440,7 @@
 	if(has_gravity)
 		clear_alert("weightless")
 	else
-		throw_alert("weightless", /obj/screen/alert/weightless)
+		throw_alert("weightless", /atom/movable/screen/alert/weightless)
 
 // Tries to turn off things that let you see through walls, like mesons.
 // Each mob does vision a bit differently so this is just for inheritence and also so overrided procs can make the vision apply instantly if they call `..()`.
@@ -1445,7 +1451,7 @@
  * Small helper component to manage the character setup HUD icon
  */
 /datum/component/character_setup
-	var/obj/screen/character_setup/screen_icon
+	var/atom/movable/screen/character_setup/screen_icon
 
 /datum/component/character_setup/Initialize()
 	if(!ismob(parent))
@@ -1466,6 +1472,8 @@
 	if(screen_icon)
 		owner?.client?.screen -= screen_icon
 		UnregisterSignal(screen_icon, COMSIG_CLICK)
+		var/datum/hud/HUD = owner?.hud_used
+		LAZYREMOVE(HUD?.other_important, screen_icon)
 		QDEL_NULL(screen_icon)
 
 /datum/component/character_setup/proc/create_mob_button(mob/user)
@@ -1481,6 +1489,8 @@
 		screen_icon.icon = HUD.ui_style
 		screen_icon.color = HUD.ui_color
 		screen_icon.alpha = HUD.ui_alpha
+	if(isAI(user))
+		screen_icon.screen_loc = ui_ai_pda_send
 	LAZYADD(HUD.other_important, screen_icon)
 	user.client?.screen += screen_icon
 
@@ -1493,7 +1503,7 @@
 /**
  * Screen object for vore panel
  */
-/obj/screen/character_setup
+/atom/movable/screen/character_setup
 	name = "character setup"
 	icon = 'icons/mob/screen/midnight.dmi'
 	icon_state = "character"
